restrucure content and add xml file for each new table
