+++ .gitignore
@@ -1,2 +1,4 @@
 __init__.py
 main.py
+ui/templetes/Library-Management-System/run
+ui/templetes/Library-Management-System/README
